Document the limits of serial transmission asyncronicity
Serial.write(), Serial.print(), Serial.println() are asynchronous only if there is enough free space in the transmit buffer.

Since write() is the base function for transmit operations, I moved the documentation to that page and linked to it from Serial.print() and Serial.println(). Since this is advanced information, I think this approach was worthwhile in the interest of avoiding duplicate content, even at the expense of making this specific part of the documentation slightly less user friendly due to the necessity of following a link. I think this may actually make the documentation more beginner friendly because it hides the information that is of no interest to a beginner in a reference page that they are less likely to read, while still making it quite accessible if they actually do want to know.